requirements.txt
fastapi
pypdf
tiktoken
sentence-transformers
faiss-cpu
